pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ronin-jobs"
version = "2.0.0"
description = "AI-Powered Job Application Automation"
authors = [{name = "AutomationChad"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
]
dependencies = [
    "requests>=2.26.0",
    "beautifulsoup4>=4.9.3",
    "openai>=1.0.0,<2.0.0",
    "anthropic>=0.39.0",
    "google-auth>=2.30.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.170.0",
    "python-dotenv>=0.19.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.0.0",
    "loguru>=0.7.2",
    "rich>=13.0.0",
    "textual>=0.80.0",
    "selenium>=4.15.2",
    "webdriver-manager>=4.0.1",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.4.0",
    "flask>=3.0.0",
    "apscheduler>=3.10.0",
    "tabulate>=0.9.0",
    "python-Levenshtein>=0.25.0",
    "nltk>=3.8.0",
    "psycopg[binary]>=3.2.0",
    "filelock>=3.12.0",
    "platformdirs>=4.0.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "black==24.2.0",
    "flake8>=7.0.0",
    "pre-commit>=3.0.0",
    "isort>=5.12.0",
    "pytest>=7.0.0",
]

[project.scripts]
ronin = "ronin.cli.main:main"

[project.urls]
Homepage = "https://github.com/automationchad/ronin"
Repository = "https://github.com/automationchad/ronin"
Issues = "https://github.com/automationchad/ronin/issues"

[tool.setuptools.packages.find]
include = ["ronin*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
